fix: add error handling for custom document

+++ packages/makeswift/src/integrate-next-app.ts
@@ -69,6 +69,14 @@ function addMakeswiftPages({ dir }: { dir: string }): void {
   }
 
   // custom document
+  if (
+    fs.existsSync(path.join(pagesFolder, `_document.${extension}`)) ||
+    fs.existsSync(path.join(pagesFolder, `_document.${extension}x`))
+  ) {
+    throw new Error(
+      'A custom document already exists, you will have to manually integrate: https://www.makeswift.com/docs/guides/manual-setup#set-up-custom-document',
+    )
+  }
   const customDocument = `export { Document as default } from '@makeswift/runtime/next'`
   fs.writeFileSync(path.join(pagesFolder, `_document.${extension}`), customDocument)
 }
